+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Faucet now supports the usage of a remote transaction prover (#830).
 - [BUGFIX] Prevents duplicated note IDs (#842).
 - Added a required Proof-of-Work in the faucet to request tokens (#831).
+- Added an optional API key request parameter to skip PoW in faucet (#839).
 
 ## v0.8.2 (2025-05-04)
 
@@ -30,7 +31,7 @@
 ### Fixes
 
 - Faucet times out during high load (#819).
- 
+
 ## v0.8.0 (2025-03-26)
 
 ### Enhancements